feat: flatten node enum
- instead of separating the Node enum into Branch and Leaf nested enums,
Node now contains all the elements it's expected to hold immediately.

-- rationale:

Did this because it can make recognizing a certain element much more verbose
since you first need to destructure the Node into the subenum, and adds
an uncessary level of indirection. it just makes the structure harder to
work with, for a hypothetical user and within the parser too.

(note how some verbosity was able to removed by forgoing the nested enum
in markup.rs & link.rs)

- Wrapped large "Leaf" variants in boxes

Prior to this change, the entire Node was 72 bytes, even though a lot of
the "Branch" nodes were wrapped in pointers. this is due to a few
partiucarly large structs blowing up its size, and losing the potential
space benefit from the Rc (although this is not the primary reason for
Rc). To minimize size, structs with more than one field were wrapped
in a Box, so they're allocated on the heap. Now I think Node should be
24 bytes (barring alignment).

- BlankLine and Softbreak created to be ZST's.
We still want them wrapped in a Match, but their contents are not
trakced

- Added the skeleton for PlainList and BlankLine

chore: cleanup comments