refactor: improve Match and stop using to_str
- made word not return a Match, just an index.
word is generally only used when you're trying to match a specific sequence you
already know. if the size is preknown, then all you care about is what
the result of the match was.

now it just returns how much the "cursor" was advanced

- a while ago i removed the generic param from Match, just turning into
a range object. this was a bad idea since i ended up just manually
recreating the functionality that the generic gave.
- now i can let specific functions do more work, and just return what
they need to do, instead of bouncing back to the parent to handle.
better separation of concerns with being able to just return an enum you
expect, or fail.

the key helper:

fn_until, now returns a Match<&str>

and several auxillary functions now return more specified Match's, where
i was manually implementing the Match by returning a tuple...

- cleaned up tags in headings this way.