halt progress
we can't set parents of the node if we return Match<expr>'s
because to set a parent of the child, the node needs to exist first

plus we don't really get much out of returning Match<Expr>, if
something's being returned then it's already a valid thing, and is going
in the pool. sure if we wanted to apply modifications, but.. we do that
in one place. i don't think it's a thing enough for it to be used

also markupend should be an err!