feat: implement caching
was stalling on this until i felt comfortable enough with the impl.

brings our expected runtime from exponential to linear. this is reflected by the `beeg` test, which saw
```
now:

        PASS [   0.003s] org-exporter tests::test_beeg

before:

        PASS [   2.831s] org-exporter tests::test_beeg
```
its runtime drop by several orders of magnitude.

caching prevents insane amount of reparsing since the program was designed with parsing in mind. saves on both runtime and memory, since allocated nodes aren't dropped until the end of the program ;p

* add types and stuff

* use proper cursor for headings

don't make a new cursor and start it at 0, that messes up the caching

* drop TODO from caching in readme