refactor(exp): pass buffer around export_rec
move buf out of our `Exporter` trait, and instead pass it around our
recursive exporter. this enables us to
A. not do the weird dance of write!(self) and implementing Write for the
exporter (this might not be weird, but it's a litle odd,,,)
B. pass around custom buffers within the exporter while culminating the
final output. useful for table export which relies on information about
all table cells for neatly generated output

drawbacks:
have to do `format!(...)` on everything that needs it. a little unweidly
and could probably be resolved with some macro magic. but :shrug: