feat: add affiliated keyword support
- support `#+caption`, `#+name` and `#attr_BACKEND`
- fixes off-by-one in heading title/id
- use appropriate html tag for descriptive lists
- new Expr::Affiliated object and respective enum

the following is valid:
```
|table

[[cool]]
```
(for any element, not just tables)

------

* use macro invocation for tags

* fix: bugs in list and heading

- lists weren't respecting descriptive items
- headings were trimming the title and id, re-add the +1

* require block type in certain macro invocations

where the writes are explicit, the tag_form expects the input to be a
block
it's pretty janky to have the syntatic appearnce affect the meaning,
but... it helps remove some redundant code. and it's harder to confuse
with the expr form, where the write is done within the call itself

* - allow adding variable tags

- make node_obj optional

(only elements and targets/radio-targets) can receive tags

- add special syntax for adding classes

* minor formatting fix

* drop macro approach in html.rs

too complicated and didn't help abstract much

requires reading a fairly complicated macro call to start understanding
what's happening, for no good reason

* fix broken macros

wasn't importing dollar so the match statements were busted

* don't unreachable! in macro_handle

* allow multiple affilaited keywords

add attrs field to node to express adddditoional attr to nodes

a little wasteful since most of the time there are.. no attrs,

maybe move to exporting backend?

* support arbitrary properties

use the self.prop method to add properties from the node object into the
written node

* shorten plainlist tag writing

* feat: suport captions properly

* much more expressive attr_BACKEND syntax

,#+attr_html: :class one :class one two three four :attr :attr1
|are|all|valid!

* add test for keyword behaviour

* more exprssive attr_BACKEND

* don't make attrs an option

makes it easier to work with and use

* revert overwritten strings