fix(footnotes): fix random footnote ordering
was using a BTreeMap (which is ordered), but I incorrectly assumed that
it would somehow work on the values, which is true. the values represent
the actual footnote numbers. so, switched to a vector that holds the
entries in insertion order and a hashset of names. insertion order is
the order footnotes are supposed to appear, so this is good.

- the * Footnotes heading applies to all levels of headings

- fix some bugs in the org footnote export