feat(site): use codemirror for text and webpack to bundle
using codemirror instead of a basic textarea gives us tons of extra features and makes it a lot nicer to work with:

- vim mode
- line numbers
- highlighted selection
- highlited current line
- themes (soon)

using a bundler has its upsides too:
- streamlines the build process a bit (just run one command0
- makes it more straightforward to minimize and optimize the generated bundle (1.2mb -> 450kb)
- also inlines the org source files for the different views.

it's also pretty much required to use codemirror (and any other npm package), so I don't really have a choice in the matter. The learning curve isn't as bad as it seems. 

commits:
----


* use rollup

* rename editor.mjs to main.js

* switch to webpack

pros:

- native wasm support is convenient
- compatible with wasm-pack (makes it really easy to import, naturally)
- easy to apply optimizations

cons:

- can't use index.html as entry point, need to use a plugin
- config is rather verbose (but descriptive!)

* move out out of inner directory

* update setup docs