perf: use iterators for template processing
before, we used to have to modify the index (+ store it in the struct so
it can be "global") which was a little hard to navigate (am i modifying
the index in the right place?), so there were a bunch of self.i+=1's
scattered around.

however, we only ever parse/look at the current element, not needing to jump around the
capture array. so, this was better implemented as a consuming iterator.
this way when an element is consumed, we know that we never need to look
at it again, and this is true for all recursive subcalls.

also saves on memory due to not needing to store all captures in memory
simultaneously.

summary: nicer code + not paying up front memory cost

there was also a bit of restructuring to cleanup the code