refactor: make ExportInner infalliable
since we now coalesce errors and don't use the ? syntax, the only other
source of those were in write!() calls. we almost exclusively write! to
in memory string buffers (or maybe stdout), so I expect write! to never
fail (at least not without the entire program catastrophically failing
right alongside it)

aside:
I guess someone could pass in some network thing as a buffer where
failure is more likely but it's not like we're recovering from a failure
anyways.

all that being said, I decided to use unwrap to just have the program
explode if write! fails. it's more appropriate than failing silently
since clearly something disastrous has happened and a panic would be
what I want in that scenario. it feels icky to cause a panic but I think
it's what a programmer would want in that scenario (and it'll almost
never happen)